+++ hunit-dejafu/CHANGELOG.rst
@@ -7,6 +7,24 @@ standard Haskell versioning scheme.
 .. _PVP: https://pvp.haskell.org/
 
 
+1.2.0.0 - No More 7.10 (2018-03-28)
+-----------------------------------
+
+* Git: :tag:`hunit-dejafu-1.2.0.0`
+* Hackage: :hackage:`hunit-dejafu-1.2.0.0`
+
+Miscellaneous
+~~~~~~~~~~~~~
+
+* GHC 7.10 support is dropped.  Dependency lower bounds are:
+
+    * :hackage:`base`: 4.9
+    * :hackage:`dejafu`: 1.5
+    * :hackage:`HUnit`: 1.3.1
+
+* The upper bound on :hackage:`dejafu` is 1.6.
+
+
 1.1.0.3 (2018-03-17)
 --------------------