+++ tasty-dejafu/CHANGELOG.rst
@@ -7,6 +7,23 @@ standard Haskell versioning scheme.
 .. _PVP: https://pvp.haskell.org/
 
 
+1.2.0.0 - No More 7.10 (2018-03-28)
+-----------------------------------
+
+* Git: :tag:`tasty-dejafu-1.2.0.0`
+* Hackage: :hackage:`tasty-dejafu-1.2.0.0`
+
+Miscellaneous
+~~~~~~~~~~~~~
+
+* GHC 7.10 support is dropped.  Dependency lower bounds are:
+
+    * :hackage:`base`: 4.9
+    * :hackage:`dejafu`: 1.5
+
+* The upper bound on :hackage:`dejafu` is 1.6.
+
+
 1.1.0.2 (2018-03-17)
 --------------------